server/requirements.txt
asgiref==3.8.1
attrs==23.2.1.dev0
autocommand==2.2.2
black==25.1.0
Brlapi==0.8.6
btrfsutil==6.13
build==1.2.2
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Cython==3.0.12
distro-info==1.10
Django==5.1.6
django-cors-headers==4.7.0
django-rest-registration==0.9.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
editables==0.5
fastjsonschema==2.21.1
filelock==3.17.0
hatchling==1.27.0
hyprshade==3.2.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
installer==0.7.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
lark==1.2.2
libfdt==1.7.2
libtorrent==2.0.11
libvirt-python==11.1.0
lit==19.1.7.dev0
louis==3.32.0
Mako==1.3.9.dev0
Markdown==3.7
MarkupSafe==2.1.5
meson==1.7.0
more-itertools==10.5.0
mypy-extensions==1.0.0
nftables==0.1
numpy==2.2.3
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
poetry-core==2.1.1
psycopg2-binary==2.9.10
pycairo==1.27.0
pycparser==2.22
PyGObject==3.50.0
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-distutils-extra==2.39
python-dotenv==1.0.1
pywal==3.3.0
Reflector==2023.6.28.0.36.1
requests==2.32.3
screeninfo==0.8.1
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
trove-classifiers==2025.1.15.22
typing_extensions==4.12.2
urllib3==2.3.0
waypaper==2.5
wheel==0.45.0
